Update Resources.resw in hungarian translation
Hey, native hungarian here! I haven't fixed everything, but made quite a few changes here and there in both grammar and explanation. If you have any questions hit me up! Fluent in english and in hungarian. (P.s. In hungarian we differentiate formal and informal, so let me know which you want to go with and I'll replace. Current form is more formal than informal, left it as is for now.) Really liked the trial of the app, couldn't buy yet tho amazing work so just wanted to help out if I could!